+++ src/js/storage.js
@@ -37,7 +37,9 @@ RUR.storage.append_world_name = function (name){
     RUR.storage.appending_world_name_flag = true;
     RUR.world_select.append_world({url:url, shortname:name, local_storage:true});
     RUR.world_select.set_url(url);  // reload as updating select choices blanks the world.
-
+    if (RUR.state.session_initialized) {
+        return;
+    }
     /* appends name to world selector and to list of possible worlds to delete */
     $('#delete-world h3').append(
         '<button class="blue-gradient inline-block" onclick="RUR.storage.delete_world(' +
